Move from scala.idioms to scala.workflow
