+++ core/src/test/scala/scala/workflow/ReadmeSpec.scala
@@ -1,4 +1,4 @@
-package scala.idioms
+package scala.workflow
 
 import language.higherKinds
 import language.postfixOps
